Update dns-process-vars.yml
Hi Brain,
We have an NFS server where we locate all of our qcows, it's readonly for users, when a user try to install dns "Copy vstat qcow2 image to dns directory" task tries to copy the qcow even though it's already there. Can we put a check here, so if the qcow is already there it won't copy it.